Refactor code editors to use a base class
