+++ src/main/java/dev/railroadide/railroad/ide/DefaultDocumentEditorStateService.java
@@ -1,6 +1,6 @@
 package dev.railroadide.railroad.ide;
 
-import dev.railroadide.railroad.ide.ui.TextEditorPane;
+import dev.railroadide.railroad.ide.ui.codeeditor.TextEditorPane;
 import dev.railroadide.railroad.plugin.spi.services.DocumentEditorStateService;
 import dev.railroadide.railroad.plugin.spi.state.Cursor;
 import dev.railroadide.railroad.plugin.spi.state.Selection;
