+++ src/main/java/dev/railroadide/railroad/ide/language/impl/JsonLanguageSupport.java
@@ -2,7 +2,9 @@ package dev.railroadide.railroad.ide.language.impl;
 
 import dev.railroadide.railroad.ide.language.BaseLanguageSupport;
 import dev.railroadide.railroad.ide.language.EditorOpenView;
+import dev.railroadide.railroad.ide.syntaxhighlighting.JsonSyntaxHighlighting;
 import dev.railroadide.railroad.ide.ui.JsonCodeEditorPane;
+import dev.railroadide.railroad.ide.ui.codeeditor.CodeEditorConfig;
 import dev.railroadide.railroad.plugin.spi.dto.Project;
 
 import java.nio.file.Path;
@@ -15,7 +17,11 @@ public final class JsonLanguageSupport extends BaseLanguageSupport {
 
     @Override
     public EditorOpenView open(Project project, Path file) {
-        var editor = new JsonCodeEditorPane(file);
+        var editor = new JsonCodeEditorPane(
+            project,
+            file,
+            CodeEditorConfig.fromLanguageSupport(project, file, this, JsonSyntaxHighlighting::computeHighlighting)
+        );
         return new EditorOpenView(
             editor,
             editor,
