+++ src/main/java/dev/railroadide/railroad/ide/ui/codeeditor/SyntaxHighlightingProvider.java
@@ -1,0 +1,10 @@
+package dev.railroadide.railroad.ide.ui.codeeditor;
+
+import org.fxmisc.richtext.model.StyleSpans;
+
+import java.util.Collection;
+
+@FunctionalInterface
+public interface SyntaxHighlightingProvider {
+    StyleSpans<Collection<String>> compute(String text);
+}
